+++ firmware/eeboot.h
@@ -11,6 +11,7 @@
 #ifdef	__cplusplus
 extern "C" {
 #endif
+    #include <stdbool.h>
     #include "eeboot_platform.h"
     
     typedef struct {
@@ -52,8 +53,8 @@ extern "C" {
     } eeboot_crc32_e;
     
     eeboot_ram_func bool eeboot_loadImage(uint32_t startAddress);
-    eeboot_ram_func bool eeboot_seekReadData(uint32_t offset, size_t numBytes, void* dataBuffer);
-    eeboot_ram_func uint32_t eeboot_getCRCseekRead(uint32_t crc32StartValue, uint32_t crc32Poly, uint32_t offset, size_t numBytes);
+    eeboot_ram_func bool eeboot_seekReadData(uint32_t offset, uint32_t numBytes, void* dataBuffer);
+    eeboot_ram_func uint32_t eeboot_getCRCseekRead(uint32_t crc32StartValue, uint32_t crc32Poly, uint32_t offset, uint32_t numBytes);
     eeboot_ram_func bool eeboot_isImageCompatible(eeboot_bootfileFixData_t *compatData);
     eeboot_ram_func bool eeboot_storeDataSegment(eeboot_segmentDescriptor_t *dataSegment);
