+++ firmware/eeboot_source_intflash.h
@@ -16,8 +16,8 @@ extern "C" {
 #define EEBOOT_SOURCE_INTFLASH_NAMESPACE(func_name) eeboot_##func_name
 #endif
     
-    eeboot_ram_func bool EEBOOT_SOURCE_INTFLASH_NAMESPACE(seekReadData)(uint32_t offset, size_t numBytes, void* dataBuffer);
-    eeboot_ram_func uint32_t EEBOOT_SOURCE_INTFLASH_NAMESPACE(getCRCseekRead)(uint32_t crc32StartValue, uint32_t crc32Poly, uint32_t offset, size_t numBytes);
+    eeboot_ram_func bool EEBOOT_SOURCE_INTFLASH_NAMESPACE(seekReadData)(uint32_t offset, uint32_t numBytes, void* dataBuffer);
+    eeboot_ram_func uint32_t EEBOOT_SOURCE_INTFLASH_NAMESPACE(getCRCseekRead)(uint32_t crc32StartValue, uint32_t crc32Poly, uint32_t offset, uint32_t numBytes);
     
 #ifdef	__cplusplus
 }
